refactor directory structure
update tests

resolve import issues
