+++ CloudHarvestCoreTasks/tasks.py
@@ -15,17 +15,16 @@ from typing import Any, List, Literal
 
 from pymongo import MongoClient
 
-from ..dataset import WalkableDict
+from .dataset import WalkableDict
 from .base import (
     BaseDataTask,
     BaseTask,
     BaseTaskChain,
-    TaskChainException,
     TaskException,
     TaskStatusCodes
 )
 
-from ..filters import MongoFilter
+from .filters import MongoFilter
 
 logger = getLogger('harvest')
 
@@ -313,12 +312,12 @@ class DataSetTask(BaseTask):
         """
         super().__init__(*args, **kwargs)
 
-        from ..dataset import DataSet
+        from .dataset import DataSet
         self.data = data if isinstance(data, DataSet) else DataSet().add_records(data)
         self.stages = stages
         self.stage_position = 0
 
-        from filters import DataSetFilter
+        from .filters import DataSetFilter
         self.filters = DataSetFilter(self.filters)
 
     def apply_filters(self) -> 'DataSetTask':
@@ -350,7 +349,7 @@ class DataSetTask(BaseTask):
             self: Returns the instance of the HarvestRecordSetTask.
         """
 
-        from ..dataset import DataSet
+        from .dataset import DataSet
 
         for stage in self.stages:
             try:
@@ -459,13 +458,13 @@ class HarvestUpdateTask(BaseTask):
         super().__init__(*args, **kwargs)
 
         # Ensure that the task chain is a BaseHarvestTaskChain
-        from .base import BaseHarvestTaskChain
+        from base import BaseHarvestTaskChain
         if not isinstance(self.task_chain, BaseHarvestTaskChain):
             raise TaskException(self, 'HarvestTask must be used in a BaseHarvestTaskChain.')
 
         # Type hint for the task_chain attribute
         from typing import cast
-        from .base import BaseHarvestTaskChain
+        from base import BaseHarvestTaskChain
         self.task_chain = cast(BaseHarvestTaskChain, self.task_chain)
 
     def method(self) -> 'HarvestUpdateTask':
@@ -479,7 +478,7 @@ class HarvestUpdateTask(BaseTask):
         self.meta['Stages'] = []
 
         # Validate the Task can reach the required silos
-        from ..silos import get_silo
+        from .silos import get_silo
         for silo_name in (self.task_chain.destination_silo, 'harvest-core'):
             try:
                 get_silo(silo_name).connect().server_info()
@@ -596,7 +595,7 @@ class HarvestUpdateTask(BaseTask):
 
         from datetime import datetime, timezone
         from pymongo import ReplaceOne
-        from ..silos import get_silo
+        from .silos import get_silo
 
         for record in data:
             # Remove an existing MongoDb _id field if it exists. This happens if the data source is MongoDB. We don't
@@ -687,7 +686,7 @@ class HarvestUpdateTask(BaseTask):
         """
         try:
             from datetime import datetime, timezone
-            from ..silos import get_silo
+            from .silos import get_silo
 
             # Deactivate Records that were not found in this data collection operation (assumed to be inactive)
             # We filter on the following fields to ensure we don't deactivate records that are collected in other processes:
@@ -944,7 +943,7 @@ class MongoTask(BaseDataTask):
         self.collection = collection
         self.result_attribute = result_attribute
 
-        from filters import MongoFilter
+        from .filters import MongoFilter
         self.filters = MongoFilter(self.filters)
 
     def apply_filters(self) -> 'MongoTask':
