+++ CloudHarvestCoreTasks/templating.py
