+++ tests/test_tasks_base.py
@@ -1,12 +1,12 @@
 import unittest
-from CloudHarvestCorePluginManager.registry import register_objects
-from ..CloudHarvestCoreTasks.tasks import *
-
+from CloudHarvestCorePluginManager.registry import register_all
+from ..CloudHarvestCoreTasks.factories import task_chain_from_dict
+from ..CloudHarvestCoreTasks.tasks import DummyTask, WaitTask, TaskStatusCodes
 
 class BaseTestCase(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        register_objects()
+        register_all()
         super(BaseTestCase, cls).setUpClass()
 
 
@@ -44,7 +44,7 @@ class TestTaskConfiguration(BaseTestCase):
                 ]
             }
         }
-        from ..CloudHarvestCoreTasks.tasks.factories import task_chain_from_dict
+        from factories import task_chain_from_dict
         self.base_task_chain = task_chain_from_dict(template=self.task_configuration)
 
     def test_instantiate(self):
@@ -94,7 +94,7 @@ class TestBaseTask(BaseTestCase):
 
     def test_retry(self):
         # Test the retry method
-        from ..CloudHarvestCoreTasks.tasks.base import BaseTaskChain
+        from base import BaseTaskChain
         task_chain = BaseTaskChain(template={
             'name': 'test_chain',
             'description': 'This is a task_chain.',
@@ -193,7 +193,6 @@ class TestBaseHarvestTaskChain(BaseTestCase):
         # Create a test silo for the task chain
         from ..CloudHarvestCoreTasks.silos import add_silo, get_silo
         from data import MONGO_TEST_RECORDS
-        from pymongo import MongoClient
 
         add_silo(name='harvest-core',
                  engine='mongo',
@@ -265,7 +264,7 @@ class TestBaseHarvestTaskChain(BaseTestCase):
             }
         }
 
-        from ..CloudHarvestCoreTasks.tasks.factories import task_chain_from_dict
+        from factories import task_chain_from_dict
         self.base_task_chain = task_chain_from_dict(template=self.task_configuration)
 
     def tearDown(self):
@@ -278,11 +277,10 @@ class TestBaseHarvestTaskChain(BaseTestCase):
         # Test the run method of the BaseHarvestTaskChain class
         self.base_task_chain.run()
         self.assertFalse(self.base_task_chain.errors)
-        self.assertEqual(len(self.base_task_chain), 3)
+        self.assertEqual(len(self.base_task_chain), 2)
         self.assertEqual(str(str(self.base_task_chain.status)), TaskStatusCodes.complete)
         self.assertIsNotNone(self.base_task_chain[1].result[0]['Tags'])
-        self.assertGreater(self.base_task_chain.result['data']['RecordsProcessed'], 1)
-        self.assertGreater(self.base_task_chain.result['data']['RecordsReplaced'], 1)
+        self.assertEqual(len(self.base_task_chain.result['data']), 10)
 
 class TestBaseTaskChain(BaseTestCase):
     """
@@ -323,7 +321,7 @@ class TestBaseTaskChain(BaseTestCase):
                 ]
             }
         }
-        from ..CloudHarvestCoreTasks.tasks.factories import task_chain_from_dict
+        from factories import task_chain_from_dict
         self.base_task_chain = task_chain_from_dict(template=self.task_configuration)
 
     def test_init(self):
@@ -504,7 +502,7 @@ class TestBaseTaskChainOnDirective(BaseTestCase):
                 ]
             }
         }
-        from ..CloudHarvestCoreTasks.tasks.factories import task_chain_from_dict
+        from factories import task_chain_from_dict
         self.base_task_chain = task_chain_from_dict(task_chain_registered_class_name='chain', template=self.task_configuration)
 
     def test_on_directives(self):
@@ -591,7 +589,7 @@ class TestBaseTaskPool(BaseTestCase):
                 }
             ]
         }
-        from ..CloudHarvestCoreTasks.tasks.base import BaseTaskChain
+        from base import BaseTaskChain
         self.base_task_chain = BaseTaskChain(template=template)
 
     def test_pooling(self):
